Move GPUI tests to sibling file and parametrize payload coverage
Two reinforcing findings from the latest review:

- The three `augmented_panic_message_includes_scenario_name_for_*_payload`
  tests are now consolidated into one `#[rstest]` with three `#[case]`s
  (`string_payload`, `str_payload`, `opaque_payload`) covering owned
  `String`, `&'static str`, and an `Opaque` `Debug`-only fall-through
  payload. The parametrized function `augmented_panic_message_includes_
  scenario_name` constructs `ScenarioMetadata` and calls
  `GpuiHarness::augmented_panic_message` once per case, asserting the
  scenario name appears in the returned string.
- The entire `#[cfg(test)] mod tests` block migrated out of
  `crates/rstest-bdd-harness-gpui/src/gpui_harness.rs` and into a new
  sibling `crates/rstest-bdd-harness-gpui/src/gpui_harness/tests.rs`. The
  parent file now carries only `#[cfg(test)] mod tests;`. Because the
  sibling file is a submodule of `gpui_harness`, the tests retain
  privileged access to private methods (`augmented_panic_message`,
  `write_stderr_diagnostic_to`) via `use super::GpuiHarness` — no
  visibility changes were required.

With the tests relocated and parametrized, `gpui_harness.rs` is back to
305 lines (well under the 400-line budget). The
`crates/rstest-bdd-harness-gpui/src/gpui_harness.rs` exemption added in
the previous commit is removed from `scripts/rs-length-allowlist.txt`.

Gates: `make check-fmt`, `make lint` (including
`check_rs_file_lengths.py` against the trimmed allowlist), `cargo test
-p rstest-bdd-harness-gpui --lib --features native-gpui-tests` (9 lib
tests pass — 3 parametrized cases + broken-pipe writer + 2 existing
behavioural tests + 3 policy tests), and `make test` (1487 passed, 7
skipped, identical contribution to the previous commit) all exit zero.
