+++ crates/rstest-bdd-harness-gpui/src/gpui_harness.rs
@@ -302,130 +302,4 @@ impl HarnessAdapter for GpuiHarness {
 }
 
 #[cfg(test)]
-mod tests {
-    //! Unit tests for the GPUI harness adapter.
-
-    use super::GpuiHarness;
-    use rstest::{fixture, rstest};
-    use rstest_bdd_harness::{
-        HarnessAdapter, ScenarioMetadata, ScenarioRunRequest, ScenarioRunner,
-    };
-
-    #[fixture]
-    fn harness() -> GpuiHarness {
-        GpuiHarness::new()
-    }
-
-    #[rstest]
-    fn gpui_harness_runs_request(harness: GpuiHarness) {
-        let request = ScenarioRunRequest::new(
-            ScenarioMetadata::new(
-                "tests/features/simple.feature",
-                "Runs in GPUI",
-                4,
-                vec!["@ui".to_string()],
-            ),
-            ScenarioRunner::new(|_context: gpui::TestAppContext| 21 * 2),
-        );
-        let result = harness
-            .run(request)
-            .unwrap_or_else(|err| panic!("gpui harness should not fail: {err}"));
-        assert_eq!(result, 42);
-    }
-
-    #[rstest]
-    fn gpui_test_context_is_available_during_run(harness: GpuiHarness) {
-        let request = ScenarioRunRequest::new(
-            ScenarioMetadata::default(),
-            ScenarioRunner::new(|context: gpui::TestAppContext| {
-                context.test_function_name().is_none()
-            }),
-        );
-        let result = harness
-            .run(request)
-            .unwrap_or_else(|err| panic!("gpui harness should not fail: {err}"));
-        assert!(result);
-    }
-
-    /// Asserts that `augmented_panic_message` embeds the scenario name in the
-    /// returned string when the panic payload is an owned `String`.
-    #[rstest]
-    fn augmented_panic_message_includes_scenario_name_for_string_payload() {
-        let metadata = ScenarioMetadata::new(
-            "tests/features/example.feature",
-            "A string payload scenario",
-            7,
-            vec![],
-        );
-        let payload: Box<dyn std::any::Any + Send> = Box::new("step panicked".to_string());
-        let message = GpuiHarness::augmented_panic_message(payload.as_ref(), &metadata);
-        assert!(
-            message.contains("A string payload scenario"),
-            "expected scenario name in: {message}"
-        );
-    }
-
-    /// Asserts that `augmented_panic_message` embeds the scenario name in the
-    /// returned string when the panic payload is a `&'static str`.
-    #[rstest]
-    fn augmented_panic_message_includes_scenario_name_for_str_payload() {
-        let metadata = ScenarioMetadata::new(
-            "tests/features/example.feature",
-            "A str payload scenario",
-            12,
-            vec![],
-        );
-        let payload: Box<dyn std::any::Any + Send> = Box::new("step panicked");
-        let message = GpuiHarness::augmented_panic_message(payload.as_ref(), &metadata);
-        assert!(
-            message.contains("A str payload scenario"),
-            "expected scenario name in: {message}"
-        );
-    }
-
-    /// Asserts that `augmented_panic_message` embeds the scenario name in the
-    /// returned string when the panic payload is an opaque type that is
-    /// neither `String` nor `&str`.
-    #[rstest]
-    fn augmented_panic_message_includes_scenario_name_for_opaque_payload() {
-        #[derive(Debug)]
-        struct Opaque;
-        let metadata = ScenarioMetadata::new(
-            "tests/features/example.feature",
-            "An opaque payload scenario",
-            17,
-            vec![],
-        );
-        let payload: Box<dyn std::any::Any + Send> = Box::new(Opaque);
-        let message = GpuiHarness::augmented_panic_message(payload.as_ref(), &metadata);
-        assert!(
-            message.contains("An opaque payload scenario"),
-            "expected scenario name in: {message}"
-        );
-    }
-
-    /// Asserts that `write_stderr_diagnostic_to` returns an `Err` and does
-    /// not panic when the underlying writer always fails with `BrokenPipe`.
-    #[rstest]
-    fn write_stderr_diagnostic_to_returns_err_on_broken_pipe() {
-        use std::io::{self, Write};
-
-        /// A `Write` implementation that always returns a `BrokenPipe` error.
-        struct BrokenPipeWriter;
-
-        impl Write for BrokenPipeWriter {
-            /// Always fails with `BrokenPipe` so the caller observes an
-            /// `Err` instead of writing any bytes.
-            fn write(&mut self, _buf: &[u8]) -> io::Result<usize> {
-                Err(io::Error::from(io::ErrorKind::BrokenPipe))
-            }
-            /// Succeeds without flushing because no bytes were written.
-            fn flush(&mut self) -> io::Result<()> {
-                Ok(())
-            }
-        }
-
-        let result = GpuiHarness::write_stderr_diagnostic_to(&mut BrokenPipeWriter, "test message");
-        assert!(result.is_err(), "expected Err from broken writer, got Ok");
-    }
-}
+mod tests;
